olivander-proto/backend/requirements.txt
fastapi
uvicorn
websockets
google-auth-oauthlib
python-dotenv
google-genai
requests
supabase
cryptography
python-jose[cryptography]
slowapi
pytest
httpx
